Исправил ошибку в названии пакета/приложения
