+++ app/src/androidTest/java/com/example/nuclearandroid_hw2/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.nucleadandroid_hw2
+package com.example.nuclearandroid_hw2
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.nucleadandroid_hw2", appContext.packageName)
+        assertEquals("com.example.nuclearandroid_hw2", appContext.packageName)
     }
 }
