+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.NucleadAndroid_hw2"
+        android:theme="@style/Theme.NuclearAndroid_hw2"
         tools:targetApi="31">
         <activity
             android:name=".ActivityFragment"
