--- app/src/main/java/com/example/nucleadandroid_hw2/ActivityFragment.kt
@@ -1,13 +1,0 @@
-package com.example.nucleadandroid_hw2
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-
-class ActivityFragment : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContentView(R.layout.activity_fragment)
-    }
-}
