+++ app/src/main/res/layout/activity_fragment.xml
@@ -9,7 +9,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragmentContainerView"
-        android:name="com.example.nucleadandroid_hw2.FragmentBA"
+        android:name="com.example.nuclearandroid_hw2.FragmentBA"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
