+++ app/src/main/res/values/themes.xml
@@ -1,9 +1,9 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.NucleadAndroid_hw2" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.NuclearAndroid_hw2" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.NucleadAndroid_hw2" parent="Base.Theme.NucleadAndroid_hw2" />
+    <style name="Theme.NuclearAndroid_hw2" parent="Base.Theme.NuclearAndroid_hw2" />
 </resources>
