+++ app/src/test/java/com/example/nuclearandroid_hw2/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.nucleadandroid_hw2
+package com.example.nuclearandroid_hw2
 
 import org.junit.Test
